requirements.txt
open_clip_torch
Pillow
torch
tqdm
numpy
